pyproject.toml
[project]
name = "12-Certification-Challenge"
version = "0.1.0"
description = "Decision Intelligence for Item Discoverability"
readme = "README.md"
requires-python = ">=3.11,<3.14"

dependencies = [
    "python-dotenv>=1.0.1",

    # Modern LangChain stack
    "langchain>=0.3.14",
    "langchain-community>=0.3.29",
    "langchain-openai>=0.3.33",
    "langgraph==0.6.7",

    # Vector store
    "pinecone>=7.3.0",
    "langchain-pinecone>=0.2.13",



    # Evaluation (modern RAGAS)
    "ragas==0.2.10",

    # Optional for reranking later
    "langchain-cohere==0.4.4",
    "cohere>=5.12.0,<5.13.0",

    # Utilities
    "tavily-python>=0.7.0",
    "ipykernel>=6.30.1",

    #Webserver
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
]
